requirements.txt
arxiv
feedparser
requests
openai
schedule
python-dotenv
beautifulsoup4
lxml
flask
flask-cors